requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
Jinja2
tenacity


